feat: disable curly rule
The curly rule causes more trouble than it solves. There are numerous occasions where having a single line within a for loop makes sense, but if that line has an if statement on it then you end up with a for loop, if statement, and everything else trying to fit onto a single line, which inevitably leads to destructured code and additional variables. I've encountered this enough times to know that this rule is less useful than it is painful.